+++ examples/key_rotation.rs
@@ -19,6 +19,7 @@ impl EncryptionProfile<String> for UserEmail {
     type Binding = FieldBound<Self>;
     type Codec = Utf8;
     type Keys = GlobalKeyContext;
+    type Padding = cryptbox::NoPadding;
 }
 
 fn main() -> Result<(), cryptbox::Error> {